+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # CHANGE LOG
+## 2023-07-04 - v0.0.6
+- Update `@dylanarmstrong/nodepub` version along with `images` -> `resources`
+- **FIX**: Clean story title to avoid `&`
+- **FIX**: Remove all empty attributes instead of just `img`
+
 ## 2023-07-03 - v0.0.5
-- Remove empty attributes from `img` tags
+- **FIX**: Remove empty attributes from `img` tags
 
 ## 2023-07-03 - v0.0.4
 - Use version for `-v` from `package.json` instead of hardcoded
 - Update default user agent, useful on CF protected sites
 
 ## 2023-07-03 - v0.0.3
-- Append a suffix to fanfiction.net cover url, so type is proper
-- Have default chapter title of `Chapter 1`, etc.
+- **FIX**: Append a suffix to fanfiction.net cover url, so type is proper
+- **FIX**: Have default chapter title of `Chapter 1`, etc.
 
 ## 2023-07-03 - v0.0.2
 - Update nodepub version